clean up some code
